Update how the notary stuff is built
